sshfs tests pass, but agonizingly slowly

+++ src/main/java/com/acrescrypto/zksync/fs/sshfs/SSHFile.java
@@ -2,6 +2,9 @@ package com.acrescrypto.zksync.fs.sshfs;
 
 import java.io.IOException;
 
+import com.acrescrypto.zksync.exceptions.EACCESException;
+import com.acrescrypto.zksync.exceptions.EISDIRException;
+import com.acrescrypto.zksync.exceptions.EMLINKException;
 import com.acrescrypto.zksync.exceptions.ENOENTException;
 import com.acrescrypto.zksync.fs.File;
 import com.acrescrypto.zksync.fs.Stat;
@@ -20,11 +23,16 @@ public class SSHFile extends File {
 		this.mode = mode;
 		
 		try {
-			fs.stat(path);
+			if((mode & O_NOFOLLOW) != 0 && fs.lstat(path).isSymlink()) throw new EMLINKException(path);
+			if(getStat().isDirectory()) throw new EISDIRException(path);
 		} catch(ENOENTException exc) {
 			if((mode & O_CREAT) == 0) throw exc;
 			fs.execAndCheck("touch", "\"" + fs.qualifiedPath(path) + "\"");
+			getStat();
 		}
+		
+		if((mode & O_APPEND) != 0) offset = getStat().getSize();
+		else if((mode & (O_TRUNC|O_WRONLY)) == (O_TRUNC|O_WRONLY)) truncate(0);
 	}
 
 	@Override
@@ -40,23 +48,30 @@ public class SSHFile extends File {
 
 	@Override
 	public void truncate(long size) throws IOException {
+		if((mode & O_WRONLY) == 0) throw new EACCESException(path);
 		fs.truncate(path, size);
+		this.cachedStat = null;
 	}
 	
 	@Override
-	public int read(byte[] buf, int offset, int maxLength) throws IOException {
+	public int read(byte[] buf, int bufOffset, int maxLength) throws IOException {
+		if((mode & O_RDONLY) == 0) throw new EACCESException(path);
+		if(offset >= cachedStat.getSize()) return 0;
 		byte[] data = fs.execAndCheck("dd",
 				"ibs=1 count=" + maxLength + " skip=" + offset + " if=\"" + fs.qualifiedPath(path) + "\"");
-		for(int i = 0; i < data.length; i++) buf[i] = data[i];
+		for(int i = 0; i < data.length; i++) buf[bufOffset+i] = data[i];
+		offset += data.length;
 		return data.length;
 	}
 
 	@Override
 	public void write(byte[] data) throws IOException {
+		if((mode & O_WRONLY) == 0) throw new EACCESException(path);
 		fs.execAndCheck("dd",
 				"obs=1 conv=notrunc seek=" + offset + " of=\"" + fs.qualifiedPath(path) + "\"",
 				data);
 		offset += data.length;
+		if(offset > getStat().getSize()) getStat().setSize(offset);
 	}
 
 	@Override
@@ -88,6 +103,7 @@ public class SSHFile extends File {
 	public void copy(File file) throws IOException {
 		truncate(0);
 		seek(0, SEEK_SET);
+		file.rewind();
 		
 		while(true) {
 			byte[] data = file.read(64*1024);
